pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "evacuation-agent-system"
version = "0.1.0"
description = "Evacuation QA agent system with LangChain, LangGraph and RAG."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "langchain>=0.3.0",
  "langchain-community>=0.3.0",
  "langchain-openai>=0.2.0",
  "langchain-text-splitters>=0.3.0",
  "langgraph>=0.2.0",
  "faiss-cpu>=1.8.0",
  "pydantic>=2.7.0",
  "python-dotenv>=1.0.1",
  "pypdf>=5.0.0",
  "rank-bm25>=0.2.2",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
